requirements.txt
flet==0.24
fletFlow @ git+https://github.com/ipritom/fletFlow
paho_mqtt==2.1.0
